+++ libopenmpt/dox/changelog.md
@@ -7,8 +7,6 @@ is just a high-level summary.
 
 ### libopenmpt svn
 
- *  xmp-openmpt / in_openmpt: libopenmpt_settings.dll no longer requires
-    .NET 4 to be installed.
  *  foo_openmpt: Settings are now accessable via foobar2000 advanced settings.
  *  Autotools based build now supports libunmo3. Specify --enable-unmo3.
  *  Support for dynamic loading of libunmo3 on MacOS X.
@@ -18,6 +16,8 @@ is just a high-level summary.
         (see `build/vs2008`)
  *  libopenmpt_ext.hpp is now distributed by default. The API is still
     considered experimental and not guaranteed to stay API or ABI compatible.
+ *  xmp-openmpt / in_openmpt: No more libopenmpt_settings.dll.
+    The settings dialog now uses a statically linked copy of MFC.
 
  *  [Bug] The -autotools tarballs were not working at all.